Further move of slack code to module. Cleanup of checkForNewSubmissions.
